+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * [FOLIO-4553](https://folio-org.atlassian.net/browse/FOLIO-4553) Set "permissions: contents: read" in maven.yml
 
+### folio-spring-base
+* [FOLSPRINGS-234](https://folio-org.atlassian.net/browse/FOLSPRINGS-234) Make execution context header lookups case-insensitive
+
 ## 10.0.0 2026-04-06
 
 **Breaking Changes:**